1) Changed the WEBCMD config option to WEB_CMD. This made it more consistent with other options. The old name is tsill recognised.
2) The UPDT_ON_OS_CHANGE option, which runs '--propupd' automatically if the
O/S changes, was incorrectly defaulting to 1. It now defaults to 0.

3) Moved several of the config options from within the tests to the start of
the program (there is a reason!). There is more of this to be done.

4) Improved the consistency of the config options. Those which take a 0/1
number are tested for correct values. Those taking a file/dir pathname are
checked for globbing, except where allowed, and other unwanted items such as
'..' and '/'. Allowed more of the options to be space-separated lists. There
is more of this to be done.

5) Included a libsafe check to look for the library in /lib64. Not too sure
about this, but it seems that it is perfectly possible to build a 64-bit
version of the library and install it into /lib64.

6) Ensure that the result of the libsafe test, a non-existent prelink command
test, and a non-compatible prelink hash function check are reported.