Removed the details of the 'ports' test from being displayed on the screen. We don't do this for other tests (except file properties, and rootkits), and so brings the test inline with the others. The details of the test - i.e. which port is being checked - is logged as before, although this is now dependent on verbose logging being enabled. Warnings are logged regardless of verbose logging being enabled or not of course.
Note: We could reduce the displayed detail of the file properties and
rootkit checks as well. However these are probably the 2 most commonly run
checks, and users usually want to see what is happening. So we won't do that.
For the other tests - such as ports - I doubt they are so bothered. So just logging
the details, and displaying the overall result on the screen is fine.